requirements.txt
certifi==2020.12.5
chardet==4.0.0
future==0.18.2
idna==2.10
line-bot-sdk==1.18.0
pytz==2020.5
requests==2.25.1
urllib3==1.26.2
wincertstore==0.2
